Implement block destroying when pressing B

+++ src/chunk/mesh.hpp
@@ -9,9 +9,9 @@ namespace chunk {
 
 class Chunk;
 
-constexpr size_t block_width = 32;
-constexpr size_t block_depth = 32;
-constexpr size_t block_height = 32;
+constexpr int block_width = 32;
+constexpr int block_depth = 32;
+constexpr int block_height = 32;
 
 class BlockArray {
 public:
